requirements.txt
django
gunicorn
django-heroku
discord
python-decouple